+++ lib/pdu.c
@@ -235,7 +235,7 @@ int rpc_process_pdu(struct rpc_context *rpc, char *buf, int size)
 	/* reassembly */
 	if (recordmarker != 0 && rpc->fragments != NULL) {
 		struct rpc_fragment *fragment;
-		size_t total = size - 4;
+		uint64_t total = size - 4;
 		char *ptr;
 
 		xdr_destroy(&xdr);
